Migrate Satellite Constallations to also insert into database
