+++ src/harvester/satellite-constellations/Dockerfile
@@ -1,0 +1,8 @@
+FROM node:22-bookworm-slim
+
+COPY . /app
+WORKDIR /app
+
+RUN cd /app
+
+CMD npm install && npm run start
