--- src/satellite-constellations/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "satellite-constellations",
-  "version": "1.0.0",
-  "description": "",
-  "main": "src/main.ts",
-  "scripts": {
-    "compile": "tsc",
-    "build": "npm run compile && node built/main.js"
-  },
-  "author": "Robert Richter",
-  "dependencies": {
-    "read-last-lines": "^1.8.0",
-    "typescript": "^5.4.4",
-    "undici-types": "^6.12.0"
-  },
-  "devDependencies": {
-    "@types/node": "^20.12.6"
-  }
-}
