--- src/satellite-constellations/satellite-numbers.nix
@@ -1,11 +1,0 @@
-let
-  pkgs = import ../../nixpkgs {};
-in pkgs.buildNpmPackage rec {
-  pname = "satelliteNumbers";
-  version = "0.0.1";
-
-  npmDepsHash = "sha256-NLBJ7ntKUhP9sNm1t3jMUap+Jk/sQnA9lX34DcpZJSw=";
-
-  src = ./.;
-}
-
