--- src/satellite-constellations/tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "nodenext",
-		"noImplicitAny": true,
-		"removeComments": true,
-		"preserveConstEnums": true,
-		"outDir": "built",
-		"sourceMap": true,
-		"moduleResolution": "nodenext",
-	},
-	"include": [
-		"src/**/*"
-	]
-}
